+++ projects/elderbyte/ngx-starter/src/lib/common/data/rest/public_api.ts
@@ -1,5 +1,3 @@
 
+
 export * from './rest-client';
-export * from './rest-client-list';
-export * from './rest-client-paged';
-export * from './rest-client-continuable';
